refactor(tab-manager): consolidate workspace definition and client
Merge workspace-client.ts into workspace.ts and update all import statements.
The original split supported two createWorkspace() calls (background + popup),
but now there's only one consumer (side panel), making the split unnecessary.
